+++ packages/docs/docs/getting-started.mdx
@@ -20,15 +20,15 @@ npm install @usedapp/core
 
 Below is a simple example:
 
-import {setup, CodeWrapper} from '../src/examples/setup';
+import render from '../src/examples/render';
+import CodeWrapper from '../src/examples/CodeWrapper';
 const root = 'getting-started-example';
-const {render} = setup({root});
 import {App as GettingStartedApp} from '../src/examples/GettingStarted.tsx';
 import * as GettingStartedCode from '../loader.js!../src/examples/GettingStarted.tsx'
 
 <div id={root}>Loading example...</div>
 
-<>{render(GettingStartedApp)}</>
+<>{render(GettingStartedApp, root)}</>
 
 ---
